Implemented UserListActivity and SignupActivity.
Created the SignupActivity for the user to be added to the server for future logins. Created a userListView for the activity_user_list.xml file that will contain the user's contacts that are added on the app. Added a menu resource file for activity_user_list.xml called add_menu.xml to take care of adding a new user to the userListView. Lastly, setup the ChatActivity for when the user clicks one of their contacts in the userListView.

This commit takes care of the user logging in, signing up, and adding contacts.