Implemented UserListActivity and SignupActivity and ChatActivity.
Created the SignupActivity for the user to be added to the server for future logins. Created a userListView for the activity_user_list.xml file that will contain the user's contacts that are added on the app. Added a menu resource file for activity_user_list.xml called add_menu.xml to take care of adding a new user to the userListView. Second, setup the ChatActivity for when the user clicks one of their contacts in the userListView. Lastly, created the ChatActivity for the user to send text back and forth to the targeted user.

This commit takes care of the user logging in, signing up, adding contacts, and chatting.